+++ LeisureTimeSystem/LeisureTimeSystem/Controllers/CourseController.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using AutoMapper;
+using LeisureTimeSystem.Attributes;
 using LeisureTimeSystem.Models.BidningModels;
+using LeisureTimeSystem.Models.BidningModels.Course;
 using LeisureTimeSystem.Models.ViewModels.Course;
 using LeisureTimeSystem.Services.Services;
 using Microsoft.AspNet.Identity;
@@ -20,16 +22,37 @@ namespace LeisureTimeSystem.Controllers
             this.service = new CourseService();
         }
 
+        [LeisureTimeAuthorize]
         public ActionResult Add(int organizationId)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToAddAcourse = this.service.IsAllowedToAddAcourse(currentUserId, organizationId);
+
+            if (!isAllowedToAddAcourse)
+            {
+                throw new ArgumentException();
+            }
+
             var addNewCourseViewModel = this.service.GetAddNewCourseViewModel(organizationId);
 
             return View(addNewCourseViewModel);
         }
 
         [HttpPost]
+        [LeisureTimeAuthorize]
         public ActionResult Add(AddNewCourseBindingModel model)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToAddAcourse = this.service.IsAllowedToAddAcourse(currentUserId, model.OrganizationId);
+
+            if (!isAllowedToAddAcourse)
+            {
+                throw new ArgumentException();
+            }
+
+
             if (this.ModelState.IsValid)
             {
                 this.service.AddNewCourse(model);
@@ -46,8 +69,19 @@ namespace LeisureTimeSystem.Controllers
             return View(addNewCourseViewModel);
         }
 
+        [LeisureTimeAuthorize]
         public ActionResult Edit(int courseId)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(courseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
+
             var editViewModel = this.service.GetEditCourseViewModel(courseId);
 
             return this.View(editViewModel);
@@ -56,8 +90,19 @@ namespace LeisureTimeSystem.Controllers
 
 
         [HttpPost]
+        [LeisureTimeAuthorize]
         public ActionResult Edit(EditCourseBindingModel model)
         {
+
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(model.Id, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
             if (this.ModelState.IsValid)
             {
                 this.service.EditCourse(model);
@@ -69,17 +114,36 @@ namespace LeisureTimeSystem.Controllers
             return this.View(editViewModel);
         }
 
-
+        [LeisureTimeAuthorize]
         public ActionResult Delete(int courseId)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(courseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
             var deleteCourseViewModel = this.service.GetDeleteCourseViewModel(courseId);
 
             return View(deleteCourseViewModel);
         }
 
         [HttpPost]
+        [LeisureTimeAuthorize]
         public ActionResult Delete(DeleteCourseBindingModel model)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(model.CourseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
             if (this.ModelState.IsValid)
             {
                 this.service.DeleteCourse(model);
@@ -94,27 +158,49 @@ namespace LeisureTimeSystem.Controllers
         }
 
 
-
+        [LeisureTimeAuthorize]
         public ActionResult ManageCourseApplications(int courseId)
         {
-            var allCourseApplicationVms = this.service.GetAllCourseApplicationViewModels(courseId);
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(courseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
+            var allCourseApplicationVms = this.service.GetManageCourseApplicationsViewModel(courseId);
 
             return View(allCourseApplicationVms);
         }
 
+
+
         [HttpPost]
-        public ActionResult ManageCourseApplications(ChangeStatusApplicationBindingModel model)
+        [LeisureTimeAuthorize]
+        public ActionResult ManageCourseApplications(ManageApplicationsWrapBindingModel model)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(model.Applications[0].Course.CourseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
             if (this.ModelState.IsValid)
             {
                 this.service.ChangeStatus(model);
-                return RedirectToAction("ManageCourseApplications", new { studentId = model.StudentId, courseId = model.CourseId });
+                return RedirectToAction("ManageCourseApplications", new { courseId = model.Applications.FirstOrDefault().Course.CourseId });
             }
 
-            var changeStatusViewModel = this.service.GetAllCourseApplicationViewModels(model.CourseId);
+            var allCourseApplicationVms = this.service.GetManageCourseApplicationsViewModel(model.Applications.FirstOrDefault().Course.CourseId);
 
-            return this.PartialView(changeStatusViewModel);
+            return View(allCourseApplicationVms);
         }
+
 
         public ActionResult DisciplineCourses(int disciplineId)
         {
@@ -123,6 +209,7 @@ namespace LeisureTimeSystem.Controllers
             return View(courses);
         }
 
+        [LeisureTimeAuthorize]
         public ActionResult Apply(int courseId)
         {
             string currentUserId = User.Identity.GetUserId();
@@ -133,6 +220,7 @@ namespace LeisureTimeSystem.Controllers
         }
 
         [HttpPost]
+        [LeisureTimeAuthorize]
         public ActionResult Apply([Bind(Include = "StudentId,CourseId")] ApplicationBindingModel applicationBindingModel)
         {
             if (this.ModelState.IsValid)
@@ -148,9 +236,18 @@ namespace LeisureTimeSystem.Controllers
             return this.View(applyViewModel);
         }
 
+        [LeisureTimeAuthorize]
         public ActionResult DeleteCourseApplication(int courseId)
         {
             string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(courseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
 
             var applicationToDelete = this.service.GetDeleteCourseApplicationViewModel(courseId, currentUserId);
 
@@ -158,8 +255,18 @@ namespace LeisureTimeSystem.Controllers
         }
 
         [HttpPost]
+        [LeisureTimeAuthorize]
         public ActionResult DeleteCourseApplication(DeleteCourseApplicationBindingModel model)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isAllowedToModifyCourse = this.service.IsAllowedToModifyCourse(model.CourseId, currentUserId);
+
+            if (!isAllowedToModifyCourse)
+            {
+                throw new ArgumentException();
+            }
+
             if (this.ModelState.IsValid)
             {
                 this.service.DeleteCourseApplication(model);
@@ -167,15 +274,23 @@ namespace LeisureTimeSystem.Controllers
                 return this.RedirectToAction("Details", "Profile");
             }
 
-            string currentUserId = User.Identity.GetUserId();
-
             var applicationToDelete = this.service.GetDeleteCourseApplicationViewModel(model.CourseId, currentUserId);
 
             return View(applicationToDelete);
         }
 
+        [LeisureTimeAuthorize]
         public ActionResult RenderStudentCourses(int studentId)
         {
+            string currentUserId = User.Identity.GetUserId();
+
+            bool isOwnProfile = this.service.IsOwnProfile(studentId, currentUserId);
+
+            if (!isOwnProfile)
+            {
+                throw new ArgumentException();
+            }
+
             var coursesViewModels = this.service.GetCourseViewModelsByStudent(studentId);
 
             return this.PartialView(coursesViewModels);
